+++ lib/Documents/src/DownloadedFile.php
@@ -40,6 +40,20 @@ class DownloadedFile extends File
             return '';
         }
 
+        // Remove images double extensions
+        // for compatibility with intervention-request
+        // and only keep the last one.
+        // example: my.image.jpg.webp => my_image_jpg.webp
+        $parts = explode('.', $string);
+        if (count($parts) > 2) {
+            $extension = array_pop($parts);
+            // Keep double extension for zip, gz, xz and bz
+            if (!\in_array($extension, ['zip', 'gz', 'xz', 'bz', 'bz2', '7z', 'tgz'], true)) {
+                $filename = implode('_', $parts);
+                $string = $filename.'.'.$extension;
+            }
+        }
+
         return (new UnicodeString($string))
             ->ascii()
             ->trim()